+++ src/parser/langs/javascriptparser.ts
